requirements.txt
asgiref==3.8.1
Django==4.2
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
mysqlclient==2.2.7
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2024.2
